requirements.txt
fastapi
uvicorn[standard]
supabase
pydantic
sqlalchemy
psycopg2-binary
numpy
python-dotenv